requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alembic==1.15.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
decorator==5.2.1
docopt==0.6.2
editdistance==0.8.1
einops==0.8.1
exceptiongroup==1.2.2
fastapi==0.115.12
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
funasr==1.2.6
future==1.0.0
gradio==5.23.0
gradio_client==1.8.0
greenlet==3.1.1
groovy==0.1.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface==0.0.1
huggingface-hub==0.30.2
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.10
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.4.2
julius==0.2.7
kaldiio==2.18.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
lightning==2.5.1
lightning-utilities==0.14.2
llvmlite==0.44.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
modelscope==1.24.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
munkres==1.1.4
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
omegaconf==2.3.0
optuna==4.2.1
orjson==3.10.16
oss2==2.19.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
primePy==1.3
propcache==0.3.0
protobuf==6.30.1
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyAudioAnalysis==0.3.14
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-multipart==0.0.20
pytorch-lightning==2.5.1
pytorch-metric-learning==2.8.1
pytorch-wpe==0.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruff==0.11.2
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
semver==3.0.4
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==0.5.0.post1
speechbrain==1.0.2
SQLAlchemy==2.0.39
starlette==0.46.1
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.2.2
text-unidecode==1.3
TextGrid==1.6.1
threadpoolctl==3.6.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.6.0+cu118
torch-audiomentations==0.12.0
torch-complex==0.4.4
torch_pitch_shift==1.2.5
torchaudio==2.6.0+cu118
torchmetrics==1.7.0
torchvision==0.21.0+cu118
tqdm==4.67.1
transformers==4.51.2
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.2
umap-learn==0.5.7
unicodecsv==0.14.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
yarl==1.18.3
